split rest api part 4
