+++ src/components/app-top-bar.js
@@ -28,7 +28,7 @@ import {
     PARAM_USE_NAME,
 } from '../utils/config-params';
 import { useDispatch, useSelector } from 'react-redux';
-import { fetchAppsAndUrls, fetchNetworkElementInfos } from '../utils/rest-api';
+import { fetchAppsAndUrls } from '../utils/rest-api';
 import makeStyles from '@mui/styles/makeStyles';
 import PropTypes from 'prop-types';
 import {
@@ -65,6 +65,7 @@ import {
 } from './utils/equipment-types';
 import { ComputingType } from './computing-status/computing-type';
 import { RunningStatus } from './utils/running-status';
+import { fetchNetworkElementInfos } from '../services/study/network';
 
 const useStyles = makeStyles((theme) => ({
     tabs: {
