+++ src/components/diagrams/diagram-pane.js
@@ -15,12 +15,7 @@ import {
     PARAM_SUBSTATION_LAYOUT,
     PARAM_USE_NAME,
 } from '../../utils/config-params';
-import {
-    fetchSvg,
-    getSubstationSingleLineDiagram,
-    getVoltageLevelSingleLineDiagram,
-    getNetworkAreaDiagramUrl,
-} from '../../utils/rest-api';
+import { fetchSvg, getNetworkAreaDiagramUrl } from '../../utils/rest-api';
 import PropTypes from 'prop-types';
 import { Chip, Stack } from '@mui/material';
 import ArrowUpwardIcon from '@mui/icons-material/ArrowUpward';
@@ -56,6 +51,10 @@ import NetworkAreaDiagramContent from './networkAreaDiagram/network-area-diagram
 import { useSnackMessage } from '@gridsuite/commons-ui';
 import { setNetworkAreaDiagramNbVoltageLevels } from '../../redux/actions';
 import { useIntl } from 'react-intl';
+import {
+    getSubstationSingleLineDiagram,
+    getVoltageLevelSingleLineDiagram,
+} from '../../services/study/network';
 
 // Returns a callback that returns a promise
 const useDisplayView = (studyUuid, currentNode) => {
