+++ src/components/dialogs/equipment-id/equipment-id-selector.js
@@ -6,7 +6,6 @@
  */
 
 import React, { useEffect, useState } from 'react';
-import { fetchEquipmentsIds } from '../../../utils/rest-api';
 import { filledTextField, gridItem } from '../dialogUtils';
 import { Autocomplete, TextField } from '@mui/material';
 import { FieldLabel } from '../../utils/inputs/hooks-helpers';
@@ -17,6 +16,7 @@ import makeStyles from '@mui/styles/makeStyles';
 import { Box } from '@mui/system';
 import { FormattedMessage } from 'react-intl';
 import clsx from 'clsx';
+import { fetchEquipmentsIds } from '../../../services/study/network-map';
 
 const useStyles = makeStyles((theme) => ({
     message: {
