+++ src/components/dialogs/form-search-copy-hook.js
@@ -7,9 +7,9 @@
 
 import { useIntl } from 'react-intl';
 import { useState } from 'react';
-import { fetchNetworkElementInfos } from '../../utils/rest-api';
 import { useSnackMessage } from '@gridsuite/commons-ui';
 import { EQUIPMENT_INFOS_TYPES } from '../utils/equipment-types';
+import { fetchNetworkElementInfos } from '../../services/study/network';
 
 export const useFormSearchCopy = ({
     studyUuid,
