+++ src/components/dialogs/network-modifications/delete-voltage-level-on-line/delete-voltage-level-on-line-form.js
@@ -16,8 +16,8 @@ import {
     REPLACING_LINE_1_ID,
     REPLACING_LINE_1_NAME,
 } from 'components/utils/field-constants';
-import { fetchEquipmentsIds } from 'utils/rest-api';
 import { areIdsEqual, getObjectId } from 'components/utils/utils';
+import { fetchEquipmentsIds } from '../../../../services/study/network-map';
 
 const DeleteVoltageLevelOnLineForm = ({ studyUuid, currentNode }) => {
     const [linesOptions, setLinesOptions] = useState([]);
