+++ src/components/dialogs/network-modifications/equipment-deletion/hvdc-lcc-deletion/hvdc-lcc-deletion-utils.js
@@ -15,7 +15,8 @@ import {
 import { useCallback } from 'react';
 import { useFieldArray, useFormContext } from 'react-hook-form';
 import { useSnackMessage } from '@gridsuite/commons-ui';
-import { fetchHvdcLineWithShuntCompensators } from '../../../../../utils/rest-api';
+
+import { fetchHvdcLineWithShuntCompensators } from '../../../../../services/study/network-map';
 
 const useHvdcLccDeletion = () => {
     const { replace: replaceMcsList1 } = useFieldArray({
