+++ src/components/dialogs/network-modifications/generator/creation/generator-creation-form.js
@@ -37,7 +37,8 @@ import { ConnectivityForm } from '../../../connectivity/connectivity-form';
 import FloatInput from 'components/utils/rhf-inputs/float-input';
 import ReactiveLimitsForm from '../../../reactive-limits/reactive-limits-form';
 import SetPointsForm from '../../../set-points/set-points-form';
-import { fetchVoltageLevelsListInfos } from 'utils/rest-api';
+
+import { fetchVoltageLevelsListInfos } from '../../../../../services/study/network';
 
 const GeneratorCreationForm = ({ studyUuid, currentNode }) => {
     const [voltageLevelOptions, setVoltageLevelOptions] = useState([]);
