+++ src/components/dialogs/network-modifications/line-split-with-voltage-level/line-split-with-voltage-level-form.js
@@ -18,11 +18,11 @@ import AddIcon from '@mui/icons-material/ControlPoint';
 import EditIcon from '@mui/icons-material/Edit';
 import TextInput from 'components/utils/rhf-inputs/text-input';
 import { ConnectivityForm } from '../../connectivity/connectivity-form';
-import { fetchVoltageLevelsListInfos } from 'utils/rest-api';
 import { Button, Typography } from '@mui/material';
 import { FormattedMessage } from 'react-intl';
 import { LineToAttachOrSplitForm } from '../line-to-attach-or-split-form/line-to-attach-or-split-form';
 import VoltageLevelCreationDialog from 'components/dialogs/network-modifications/voltage-level/creation/voltage-level-creation-dialog';
+import { fetchVoltageLevelsListInfos } from '../../../../services/study/network';
 
 const LineSplitWithVoltageLevelForm = ({
     studyUuid,
