+++ src/components/dialogs/network-modifications/line/creation/line-creation-dialog.js
@@ -39,7 +39,7 @@ import { EQUIPMENT_TYPES } from 'components/utils/equipment-types';
 import PropTypes from 'prop-types';
 import React, { useCallback, useEffect, useState } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
-import { fetchVoltageLevelsListInfos, FetchStatus } from 'utils/rest-api';
+import { FetchStatus } from 'utils/rest-api';
 
 import { microUnitToUnit, unitToMicroUnit } from 'utils/rounding';
 import {
@@ -80,6 +80,7 @@ import { formatTemporaryLimits } from 'components/utils/utils';
 import LineTypeSegmentDialog from '../../../line-types-catalog/line-type-segment-dialog';
 import { useOpenShortWaitFetching } from 'components/dialogs/commons/handle-modification-form';
 import { createLine } from '../../../../../services/study/network-modifications';
+import { fetchVoltageLevelsListInfos } from '../../../../../services/study/network';
 
 const emptyFormData = {
     ...getHeaderEmptyFormData(),
