+++ src/components/dialogs/network-modifications/line/modification/line-modification-dialog.js
@@ -27,7 +27,7 @@ import {
     TOTAL_SUSCEPTANCE,
 } from 'components/utils/field-constants';
 import { FormProvider, useForm } from 'react-hook-form';
-import { fetchNetworkElementInfos, FetchStatus } from 'utils/rest-api';
+import { FetchStatus } from 'utils/rest-api';
 import { sanitizeString } from 'components/dialogs/dialogUtils';
 import { microUnitToUnit, unitToMicroUnit } from 'utils/rounding';
 import yup from 'components/utils/yup-config';
@@ -59,6 +59,7 @@ import {
 } from 'components/utils/equipment-types';
 import { EquipmentIdSelector } from '../../../equipment-id/equipment-id-selector';
 import { modifyLine } from '../../../../../services/study/network-modifications';
+import { fetchNetworkElementInfos } from '../../../../../services/study/network';
 export const LineCreationDialogTab = {
     CHARACTERISTICS_TAB: 0,
     LIMITS_TAB: 1,
