+++ src/components/dialogs/network-modifications/shunt-compensator/creation/shunt-compensator-creation-form.js
@@ -13,8 +13,8 @@ import { filledTextField, gridItem, GridSection } from '../../../dialogUtils';
 
 import TextInput from 'components/utils/rhf-inputs/text-input';
 import { ConnectivityForm } from '../../../connectivity/connectivity-form';
-import { fetchVoltageLevelsListInfos } from 'utils/rest-api';
 import { CharacteristicsForm } from '../characteristics-pane/characteristics-form';
+import { fetchVoltageLevelsListInfos } from '../../../../../services/study/network';
 
 const ShuntCompensatorCreationForm = ({ studyUuid, currentNode }) => {
     const [voltageLevelOptions, setVoltageLevelOptions] = useState([]);
