+++ src/components/dialogs/network-modifications/substation/modification/substation-modification-dialog.js
@@ -22,7 +22,7 @@ import {
     ADDED,
 } from 'components/utils/field-constants';
 import SubstationModificationForm from './substation-modification-form';
-import { fetchNetworkElementInfos, FetchStatus } from 'utils/rest-api';
+import { FetchStatus } from 'utils/rest-api';
 import { sanitizeString } from '../../../dialogUtils';
 import { useOpenShortWaitFetching } from 'components/dialogs/commons/handle-modification-form';
 import { FORM_LOADING_DELAY } from 'components/network/constants';
@@ -32,6 +32,7 @@ import {
 } from 'components/utils/equipment-types';
 import { EquipmentIdSelector } from '../../../equipment-id/equipment-id-selector';
 import { modifySubstation } from '../../../../../services/study/network-modifications';
+import { fetchNetworkElementInfos } from '../../../../../services/study/network';
 
 const checkUniquePropertiesNames = (properties) => {
     const validValues = properties.filter((v) => v?.name);
