+++ src/components/dialogs/regulating-terminal/regulating-terminal-form.js
@@ -18,8 +18,8 @@ import {
 import PropTypes from 'prop-types';
 import React, { useCallback, useEffect, useState } from 'react';
 import { useFormContext, useWatch } from 'react-hook-form';
-import { fetchVoltageLevelEquipments } from 'utils/rest-api';
 import AutocompleteInput from '../../utils/rhf-inputs/autocomplete-input';
+import { fetchVoltageLevelEquipments } from '../../../services/study/network-map';
 
 // Factory used to create a filter method that is used to change the default
 // option filter behaviour of the Autocomplete component
