+++ src/components/graph/menus/node-editor.js
@@ -10,12 +10,12 @@ import PropTypes from 'prop-types';
 import makeStyles from '@mui/styles/makeStyles';
 import { lighten, darken } from '@mui/material/styles';
 import NetworkModificationNodeEditor from './network-modification-node-editor';
-import { updateTreeNode } from '../../../utils/rest-api';
 import { useSnackMessage } from '@gridsuite/commons-ui';
 import { EditableTitle } from './editable-title';
 import { useDispatch, useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
 import { setModificationsDrawerOpen } from '../../../redux/actions';
+import { updateTreeNode } from '../../../services/study/tree';
 
 const useStyles = makeStyles((theme) => ({
     paper: {
