+++ src/components/network-modification-tree-pane.js
@@ -7,17 +7,9 @@
 
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 import {
-    createTreeNode,
-    deleteTreeNode,
-    fetchNetworkModificationTreeNode,
     getUniqueNodeName,
     buildNode,
-    copyTreeNode,
-    cutTreeNode,
-    deleteSubtree,
     fetchNetworkModificationSubtree,
-    cutSubtree,
-    copySubtree,
 } from '../utils/rest-api';
 import {
     networkModificationTreeNodeAdded,
@@ -42,6 +34,16 @@ import makeStyles from '@mui/styles/makeStyles';
 import { DRAWER_NODE_EDITOR_WIDTH } from '../utils/UIconstants';
 import ExportDialog from './dialogs/export-dialog';
 import { BUILD_STATUS, UPDATE_TYPE } from './network/constants';
+import {
+    copySubtree,
+    copyTreeNode,
+    createTreeNode,
+    cutSubtree,
+    cutTreeNode,
+    deleteSubtree,
+    deleteTreeNode,
+    fetchNetworkModificationTreeNode,
+} from '../services/study/tree';
 
 const useStyles = makeStyles((theme) => ({
     nodeEditor: {
