+++ src/services/study/network-map.js
@@ -1,0 +1,124 @@
+/**
+ * Copyright (c) 2023, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+import { backendFetchJson, getQueryParamsList } from '../../utils/rest-api';
+import { getStudyUrlWithNodeUuid } from './index';
+
+export function fetchHvdcLineWithShuntCompensators(
+    studyUuid,
+    currentNodeUuid,
+    hvdcLineId
+) {
+    console.info(
+        `Fetching HVDC Line '${hvdcLineId}' with Shunt Compensators of study '${studyUuid}' and node '${currentNodeUuid}'...`
+    );
+    const fetchEquipmentsUrl =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/network-map' +
+        '/hvdc-lines/' +
+        hvdcLineId +
+        '/shunt-compensators';
+    console.debug(fetchEquipmentsUrl);
+    return backendFetchJson(fetchEquipmentsUrl);
+}
+
+export function fetchAllEquipments(studyUuid, currentNodeUuid, substationsIds) {
+    console.info(
+        `Fetching all equipments of study '${studyUuid}' and node '${currentNodeUuid}' with substations ids '${substationsIds}'...`
+    );
+
+    const fetchEquipmentsUrl =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/network-map/all' +
+        '?' +
+        getQueryParamsList(substationsIds, 'substationId');
+    console.debug(fetchEquipmentsUrl);
+    return backendFetchJson(fetchEquipmentsUrl);
+}
+
+export function fetchVoltageLevelEquipments(
+    studyUuid,
+    currentNodeUuid,
+    substationsIds,
+    voltageLevelId,
+    inUpstreamBuiltParentNode
+) {
+    console.info(
+        `Fetching equipments of study '${studyUuid}' and node '${currentNodeUuid}' and voltage level '${voltageLevelId}' with substations ids '${substationsIds}'...`
+    );
+    let urlSearchParams = new URLSearchParams();
+    if (inUpstreamBuiltParentNode !== undefined) {
+        urlSearchParams.append(
+            'inUpstreamBuiltParentNode',
+            inUpstreamBuiltParentNode
+        );
+    }
+
+    const fetchEquipmentsUrl =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/network-map' +
+        '/voltage-level-equipments/' +
+        encodeURIComponent(voltageLevelId) +
+        '?' +
+        getQueryParamsList(substationsIds, 'substationId') +
+        urlSearchParams.toString();
+    console.debug(fetchEquipmentsUrl);
+    return backendFetchJson(fetchEquipmentsUrl);
+}
+
+export function fetchEquipmentsIds(
+    studyUuid,
+    currentNodeUuid,
+    substationsIds,
+    equipmentType,
+    inUpstreamBuiltParentNode
+) {
+    console.info(
+        `Fetching equipments ids '${equipmentType}' of study '${studyUuid}' and node '${currentNodeUuid}' with substations ids '${substationsIds}'...`
+    );
+    let urlSearchParams = new URLSearchParams();
+
+    let fetchEquipmentsUrl =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/network-map/' +
+        'equipments-ids' +
+        '?' +
+        'equipmentType=' +
+        equipmentType +
+        getQueryParamsList(substationsIds, 'substationId');
+
+    if (inUpstreamBuiltParentNode !== undefined) {
+        urlSearchParams.append(
+            'inUpstreamBuiltParentNode',
+            inUpstreamBuiltParentNode
+        );
+        fetchEquipmentsUrl =
+            fetchEquipmentsUrl + '&' + urlSearchParams.toString();
+    }
+    console.debug(fetchEquipmentsUrl);
+    return backendFetchJson(fetchEquipmentsUrl);
+}
+
+export function fetchLineOrTransformer(
+    studyUuid,
+    currentNodeUuid,
+    equipmentId
+) {
+    console.info(
+        `Fetching specific equipment '${equipmentId}' of type branch-or-3wt of study '${studyUuid}' and node '${currentNodeUuid}' ...`
+    );
+    let urlSearchParams = new URLSearchParams();
+    urlSearchParams.append('inUpstreamBuiltParentNode', true);
+    const fetchEquipmentInfosUrl =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/network-map/branch-or-3wt/' +
+        encodeURIComponent(equipmentId) +
+        '?' +
+        urlSearchParams.toString();
+    console.debug(fetchEquipmentInfosUrl);
+    return backendFetchJson(fetchEquipmentInfosUrl);
+}
