+++ src/services/study/network.js
@@ -1,0 +1,430 @@
+/**
+ * Copyright (c) 2023, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+import { getStudyUrlWithNodeUuid } from './index';
+import { backendFetchJson } from '../../utils/rest-api';
+import {
+    EQUIPMENT_INFOS_TYPES,
+    EQUIPMENT_TYPES,
+} from '../../components/utils/equipment-types';
+
+/* voltage-levels */
+export function getVoltageLevelSingleLineDiagram(
+    studyUuid,
+    currentNodeUuid,
+    voltageLevelId,
+    useName,
+    centerLabel,
+    diagonalLabel,
+    componentLibrary,
+    sldDisplayMode,
+    language
+) {
+    console.info(
+        `Getting url of voltage level diagram '${voltageLevelId}' of study '${studyUuid}' and node '${currentNodeUuid}'...`
+    );
+    return (
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/network/voltage-levels/' +
+        encodeURIComponent(voltageLevelId) +
+        '/svg-and-metadata?' +
+        new URLSearchParams({
+            useName: useName,
+            centerLabel: centerLabel,
+            diagonalLabel: diagonalLabel,
+            topologicalColoring: true,
+            ...(componentLibrary !== null && {
+                componentLibrary: componentLibrary,
+            }),
+            sldDisplayMode: sldDisplayMode,
+            language: language,
+        }).toString()
+    );
+}
+
+export function fetchBusesForVoltageLevel(
+    studyUuid,
+    currentNodeUuid,
+    voltageLevelId
+) {
+    console.info(
+        `Fetching buses of study '${studyUuid}' and node '${currentNodeUuid}' + ' for voltage level '${voltageLevelId}'...`
+    );
+    let urlSearchParams = new URLSearchParams();
+    urlSearchParams.append('inUpstreamBuiltParentNode', true);
+
+    const fetchBusesUrl =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/network/voltage-levels/' +
+        encodeURIComponent(voltageLevelId) +
+        '/buses' +
+        '?' +
+        urlSearchParams.toString();
+    console.debug(fetchBusesUrl);
+    return backendFetchJson(fetchBusesUrl);
+}
+
+export function fetchBusbarSectionsForVoltageLevel(
+    studyUuid,
+    currentNodeUuid,
+    voltageLevelId
+) {
+    console.info(
+        `Fetching busbar sections of study '${studyUuid}' and node '${currentNodeUuid}' + ' for voltage level '${voltageLevelId}'...`
+    );
+    let urlSearchParams = new URLSearchParams();
+    urlSearchParams.append('inUpstreamBuiltParentNode', true);
+
+    const fetchBusbarSectionsUrl =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/network/voltage-levels/' +
+        encodeURIComponent(voltageLevelId) +
+        '/busbar-sections' +
+        '?' +
+        urlSearchParams.toString();
+
+    console.debug(fetchBusbarSectionsUrl);
+    return backendFetchJson(fetchBusbarSectionsUrl);
+}
+
+/* substations */
+export function getSubstationSingleLineDiagram(
+    studyUuid,
+    currentNodeUuid,
+    substationId,
+    useName,
+    centerLabel,
+    diagonalLabel,
+    substationLayout,
+    componentLibrary,
+    language
+) {
+    console.info(
+        `Getting url of substation diagram '${substationId}' of study '${studyUuid}' and node '${currentNodeUuid}'...`
+    );
+    return (
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/network/substations/' +
+        encodeURIComponent(substationId) +
+        '/svg-and-metadata?' +
+        new URLSearchParams({
+            useName: useName,
+            centerLabel: centerLabel,
+            diagonalLabel: diagonalLabel,
+            topologicalColoring: true,
+            substationLayout: substationLayout,
+            ...(componentLibrary !== null && {
+                componentLibrary: componentLibrary,
+            }),
+            language: language,
+        }).toString()
+    );
+}
+
+/* elements */
+export function fetchNetworkElementsInfos(
+    studyUuid,
+    currentNodeUuid,
+    substationsIds,
+    elementType,
+    infoType,
+    inUpstreamBuiltParentNode
+) {
+    console.info(
+        `Fetching network '${elementType}' elements '${infoType}' infos of study '${studyUuid}' and node '${currentNodeUuid}' with substations ids '${substationsIds}'...`
+    );
+
+    let urlSearchParams = new URLSearchParams();
+    if (inUpstreamBuiltParentNode !== undefined) {
+        urlSearchParams.append(
+            'inUpstreamBuiltParentNode',
+            inUpstreamBuiltParentNode
+        );
+    }
+    if (substationsIds !== undefined && substationsIds.length > 0) {
+        substationsIds.forEach((id) =>
+            urlSearchParams.append('substationsIds', id)
+        );
+    }
+    urlSearchParams.append('elementType', elementType);
+    urlSearchParams.append('infoType', infoType);
+
+    const fetchElementsUrl =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/network/elements' +
+        '?' +
+        urlSearchParams.toString();
+    console.debug(fetchElementsUrl);
+
+    return backendFetchJson(fetchElementsUrl);
+}
+
+export function fetchNetworkElementInfos(
+    studyUuid,
+    currentNodeUuid,
+    elementType,
+    infoType,
+    elementId,
+    inUpstreamBuiltParentNode
+) {
+    console.info(
+        `Fetching specific network element '${elementId}' of type '${elementType}' of study '${studyUuid}' and node '${currentNodeUuid}' ...`
+    );
+    let urlSearchParams = new URLSearchParams();
+    if (inUpstreamBuiltParentNode !== undefined) {
+        urlSearchParams.append(
+            'inUpstreamBuiltParentNode',
+            inUpstreamBuiltParentNode
+        );
+    }
+    urlSearchParams.append('elementType', elementType);
+    urlSearchParams.append('infoType', infoType);
+
+    const fetchElementsUrl =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/network/elements/' +
+        encodeURIComponent(elementId) +
+        '?' +
+        urlSearchParams.toString();
+    console.debug(fetchElementsUrl);
+
+    return backendFetchJson(fetchElementsUrl);
+}
+
+export function fetchSubstationsMapInfos(
+    studyUuid,
+    currentNodeUuid,
+    substationsIds,
+    inUpstreamBuiltParentNode
+) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.SUBSTATION.type,
+        EQUIPMENT_INFOS_TYPES.MAP.type,
+        inUpstreamBuiltParentNode
+    );
+}
+
+export function fetchLinesMapInfos(
+    studyUuid,
+    currentNodeUuid,
+    substationsIds,
+    inUpstreamBuiltParentNode
+) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.LINE.type,
+        EQUIPMENT_INFOS_TYPES.MAP.type,
+        inUpstreamBuiltParentNode
+    );
+}
+
+export function fetchHvdcLinesMapInfos(
+    studyUuid,
+    currentNodeUuid,
+    substationsIds,
+    inUpstreamBuiltParentNode
+) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.HVDC_LINE.type,
+        EQUIPMENT_INFOS_TYPES.MAP.type,
+        inUpstreamBuiltParentNode
+    );
+}
+
+export function fetchSubstations(studyUuid, currentNodeUuid, substationsIds) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.SUBSTATION.type,
+        EQUIPMENT_INFOS_TYPES.TAB.type,
+        true
+    );
+}
+
+export function fetchLines(studyUuid, currentNodeUuid, substationsIds) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.LINE.type,
+        EQUIPMENT_INFOS_TYPES.TAB.type,
+        true
+    );
+}
+
+export function fetchVoltageLevels(studyUuid, currentNodeUuid, substationsIds) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.VOLTAGE_LEVEL.type,
+        EQUIPMENT_INFOS_TYPES.TAB.type,
+        true
+    );
+}
+
+export function fetchVoltageLevelsListInfos(
+    studyUuid,
+    currentNodeUuid,
+    substationsIds
+) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.VOLTAGE_LEVEL.type,
+        EQUIPMENT_INFOS_TYPES.LIST.type,
+        true
+    );
+}
+
+export function fetchTwoWindingsTransformers(
+    studyUuid,
+    currentNodeUuid,
+    substationsIds
+) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.TWO_WINDINGS_TRANSFORMER.type,
+        EQUIPMENT_INFOS_TYPES.TAB.type,
+        false
+    );
+}
+
+export function fetchThreeWindingsTransformers(
+    studyUuid,
+    currentNodeUuid,
+    substationsIds
+) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.THREE_WINDINGS_TRANSFORMER.type,
+        EQUIPMENT_INFOS_TYPES.TAB.type,
+        false
+    );
+}
+
+export function fetchGenerators(studyUuid, currentNodeUuid, substationsIds) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.GENERATOR.type,
+        EQUIPMENT_INFOS_TYPES.TAB.type
+    );
+}
+
+export function fetchLoads(studyUuid, currentNodeUuid, substationsIds) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.LOAD.type,
+        EQUIPMENT_INFOS_TYPES.TAB.type,
+        true
+    );
+}
+
+export function fetchDanglingLines(studyUuid, currentNodeUuid, substationsIds) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.DANGLING_LINE.type,
+        EQUIPMENT_INFOS_TYPES.TAB.type
+    );
+}
+
+export function fetchBatteries(studyUuid, currentNodeUuid, substationsIds) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.BATTERY.type,
+        EQUIPMENT_INFOS_TYPES.TAB.type
+    );
+}
+
+export function fetchHvdcLines(studyUuid, currentNodeUuid, substationsIds) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.HVDC_LINE.type,
+        EQUIPMENT_INFOS_TYPES.TAB.type,
+        true
+    );
+}
+
+export function fetchLccConverterStations(
+    studyUuid,
+    currentNodeUuid,
+    substationsIds
+) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.LCC_CONVERTER_STATION.type,
+        EQUIPMENT_INFOS_TYPES.TAB.type
+    );
+}
+
+export function fetchVscConverterStations(
+    studyUuid,
+    currentNodeUuid,
+    substationsIds
+) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.VSC_CONVERTER_STATION.type,
+        EQUIPMENT_INFOS_TYPES.TAB.type
+    );
+}
+
+export function fetchShuntCompensators(
+    studyUuid,
+    currentNodeUuid,
+    substationsIds
+) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.SHUNT_COMPENSATOR.type,
+        EQUIPMENT_INFOS_TYPES.TAB.type
+    );
+}
+
+export function fetchStaticVarCompensators(
+    studyUuid,
+    currentNodeUuid,
+    substationsIds
+) {
+    return fetchNetworkElementsInfos(
+        studyUuid,
+        currentNodeUuid,
+        substationsIds,
+        EQUIPMENT_TYPES.STATIC_VAR_COMPENSATOR.type,
+        EQUIPMENT_INFOS_TYPES.TAB.type
+    );
+}
