+++ src/utils/rest-api.js
@@ -6,10 +6,6 @@
  */
 
 import { store } from '../redux/store';
-import {
-    EQUIPMENT_INFOS_TYPES,
-    EQUIPMENT_TYPES,
-} from '../components/utils/equipment-types';
 
 const PREFIX_STUDY_QUERIES = process.env.REACT_APP_API_GATEWAY + '/study';
 export const getWsBase = () =>
@@ -160,72 +156,6 @@ export function fetchStudyExists(studyUuid) {
     return backendFetch(fetchStudiesUrl, { method: 'head' });
 }
 
-export function getVoltageLevelSingleLineDiagram(
-    studyUuid,
-    currentNodeUuid,
-    voltageLevelId,
-    useName,
-    centerLabel,
-    diagonalLabel,
-    componentLibrary,
-    sldDisplayMode,
-    language
-) {
-    console.info(
-        `Getting url of voltage level diagram '${voltageLevelId}' of study '${studyUuid}' and node '${currentNodeUuid}'...`
-    );
-    return (
-        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
-        '/network/voltage-levels/' +
-        encodeURIComponent(voltageLevelId) +
-        '/svg-and-metadata?' +
-        new URLSearchParams({
-            useName: useName,
-            centerLabel: centerLabel,
-            diagonalLabel: diagonalLabel,
-            topologicalColoring: true,
-            ...(componentLibrary !== null && {
-                componentLibrary: componentLibrary,
-            }),
-            sldDisplayMode: sldDisplayMode,
-            language: language,
-        }).toString()
-    );
-}
-
-export function getSubstationSingleLineDiagram(
-    studyUuid,
-    currentNodeUuid,
-    substationId,
-    useName,
-    centerLabel,
-    diagonalLabel,
-    substationLayout,
-    componentLibrary,
-    language
-) {
-    console.info(
-        `Getting url of substation diagram '${substationId}' of study '${studyUuid}' and node '${currentNodeUuid}'...`
-    );
-    return (
-        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
-        '/network/substations/' +
-        encodeURIComponent(substationId) +
-        '/svg-and-metadata?' +
-        new URLSearchParams({
-            useName: useName,
-            centerLabel: centerLabel,
-            diagonalLabel: diagonalLabel,
-            topologicalColoring: true,
-            substationLayout: substationLayout,
-            ...(componentLibrary !== null && {
-                componentLibrary: componentLibrary,
-            }),
-            language: language,
-        }).toString()
-    );
-}
-
 export function getNetworkAreaDiagramUrl(
     studyUuid,
     currentNodeUuid,
@@ -246,7 +176,7 @@ export function getNetworkAreaDiagramUrl(
     );
 }
 
-function getQueryParamsList(params, paramName) {
+export function getQueryParamsList(params, paramName) {
     if (params !== undefined && params.length > 0) {
         const urlSearchParams = new URLSearchParams();
         params.forEach((id) => urlSearchParams.append(paramName, id));
@@ -278,17 +208,6 @@ export function fetchSvg(svgUrl) {
     );
 }
 
-export function fetchSubstations(studyUuid, currentNodeUuid, substationsIds) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.SUBSTATION.type,
-        EQUIPMENT_INFOS_TYPES.TAB.type,
-        true
-    );
-}
-
 export function fetchSubstationPositions(
     studyUuid,
     currentNodeUuid,
@@ -309,199 +228,6 @@ export function fetchSubstationPositions(
         paramsList;
     console.debug(fetchSubstationPositionsUrl);
     return backendFetchJson(fetchSubstationPositionsUrl);
-}
-
-export function fetchLines(studyUuid, currentNodeUuid, substationsIds) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.LINE.type,
-        EQUIPMENT_INFOS_TYPES.TAB.type,
-        true
-    );
-}
-
-export function fetchVoltageLevels(studyUuid, currentNodeUuid, substationsIds) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.VOLTAGE_LEVEL.type,
-        EQUIPMENT_INFOS_TYPES.TAB.type,
-        true
-    );
-}
-
-export function fetchVoltageLevelsListInfos(
-    studyUuid,
-    currentNodeUuid,
-    substationsIds
-) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.VOLTAGE_LEVEL.type,
-        EQUIPMENT_INFOS_TYPES.LIST.type,
-        true
-    );
-}
-
-export function fetchTwoWindingsTransformers(
-    studyUuid,
-    currentNodeUuid,
-    substationsIds
-) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.TWO_WINDINGS_TRANSFORMER.type,
-        EQUIPMENT_INFOS_TYPES.TAB.type,
-        false
-    );
-}
-
-export function fetchThreeWindingsTransformers(
-    studyUuid,
-    currentNodeUuid,
-    substationsIds
-) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.THREE_WINDINGS_TRANSFORMER.type,
-        EQUIPMENT_INFOS_TYPES.TAB.type,
-        false
-    );
-}
-
-export function fetchGenerators(studyUuid, currentNodeUuid, substationsIds) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.GENERATOR.type,
-        EQUIPMENT_INFOS_TYPES.TAB.type
-    );
-}
-
-export function fetchLoads(studyUuid, currentNodeUuid, substationsIds) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.LOAD.type,
-        EQUIPMENT_INFOS_TYPES.TAB.type,
-        true
-    );
-}
-
-export function fetchDanglingLines(studyUuid, currentNodeUuid, substationsIds) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.DANGLING_LINE.type,
-        EQUIPMENT_INFOS_TYPES.TAB.type
-    );
-}
-
-export function fetchBatteries(studyUuid, currentNodeUuid, substationsIds) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.BATTERY.type,
-        EQUIPMENT_INFOS_TYPES.TAB.type
-    );
-}
-
-export function fetchHvdcLines(studyUuid, currentNodeUuid, substationsIds) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.HVDC_LINE.type,
-        EQUIPMENT_INFOS_TYPES.TAB.type,
-        true
-    );
-}
-
-export function fetchHvdcLineWithShuntCompensators(
-    studyUuid,
-    currentNodeUuid,
-    hvdcLineId
-) {
-    console.info(
-        `Fetching HVDC Line '${hvdcLineId}' with Shunt Compensators of study '${studyUuid}' and node '${currentNodeUuid}'...`
-    );
-    const fetchEquipmentsUrl =
-        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
-        '/network-map' +
-        '/hvdc-lines/' +
-        hvdcLineId +
-        '/shunt-compensators';
-    console.debug(fetchEquipmentsUrl);
-    return backendFetchJson(fetchEquipmentsUrl);
-}
-
-export function fetchLccConverterStations(
-    studyUuid,
-    currentNodeUuid,
-    substationsIds
-) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.LCC_CONVERTER_STATION.type,
-        EQUIPMENT_INFOS_TYPES.TAB.type
-    );
-}
-
-export function fetchVscConverterStations(
-    studyUuid,
-    currentNodeUuid,
-    substationsIds
-) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.VSC_CONVERTER_STATION.type,
-        EQUIPMENT_INFOS_TYPES.TAB.type
-    );
-}
-
-export function fetchShuntCompensators(
-    studyUuid,
-    currentNodeUuid,
-    substationsIds
-) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.SHUNT_COMPENSATOR.type,
-        EQUIPMENT_INFOS_TYPES.TAB.type
-    );
-}
-
-export function fetchStaticVarCompensators(
-    studyUuid,
-    currentNodeUuid,
-    substationsIds
-) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.STATIC_VAR_COMPENSATOR.type,
-        EQUIPMENT_INFOS_TYPES.TAB.type
-    );
 }
 
 export function searchEquipmentsInfos(
@@ -537,172 +263,6 @@ export function searchEquipmentsInfos(
     );
 }
 
-export function fetchAllEquipments(studyUuid, currentNodeUuid, substationsIds) {
-    console.info(
-        `Fetching all equipments of study '${studyUuid}' and node '${currentNodeUuid}' with substations ids '${substationsIds}'...`
-    );
-
-    const fetchEquipmentsUrl =
-        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
-        '/network-map/all' +
-        '?' +
-        getQueryParamsList(substationsIds, 'substationId');
-    console.debug(fetchEquipmentsUrl);
-    return backendFetchJson(fetchEquipmentsUrl);
-}
-
-export function fetchNetworkElementsInfos(
-    studyUuid,
-    currentNodeUuid,
-    substationsIds,
-    elementType,
-    infoType,
-    inUpstreamBuiltParentNode
-) {
-    console.info(
-        `Fetching network '${elementType}' elements '${infoType}' infos of study '${studyUuid}' and node '${currentNodeUuid}' with substations ids '${substationsIds}'...`
-    );
-
-    let urlSearchParams = new URLSearchParams();
-    if (inUpstreamBuiltParentNode !== undefined) {
-        urlSearchParams.append(
-            'inUpstreamBuiltParentNode',
-            inUpstreamBuiltParentNode
-        );
-    }
-    if (substationsIds !== undefined && substationsIds.length > 0) {
-        substationsIds.forEach((id) =>
-            urlSearchParams.append('substationsIds', id)
-        );
-    }
-    urlSearchParams.append('elementType', elementType);
-    urlSearchParams.append('infoType', infoType);
-
-    const fetchElementsUrl =
-        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
-        '/network/elements' +
-        '?' +
-        urlSearchParams.toString();
-    console.debug(fetchElementsUrl);
-
-    return backendFetchJson(fetchElementsUrl);
-}
-
-export function fetchNetworkElementInfos(
-    studyUuid,
-    currentNodeUuid,
-    elementType,
-    infoType,
-    elementId,
-    inUpstreamBuiltParentNode
-) {
-    console.info(
-        `Fetching specific network element '${elementId}' of type '${elementType}' of study '${studyUuid}' and node '${currentNodeUuid}' ...`
-    );
-    let urlSearchParams = new URLSearchParams();
-    if (inUpstreamBuiltParentNode !== undefined) {
-        urlSearchParams.append(
-            'inUpstreamBuiltParentNode',
-            inUpstreamBuiltParentNode
-        );
-    }
-    urlSearchParams.append('elementType', elementType);
-    urlSearchParams.append('infoType', infoType);
-
-    const fetchElementsUrl =
-        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
-        '/network/elements/' +
-        encodeURIComponent(elementId) +
-        '?' +
-        urlSearchParams.toString();
-    console.debug(fetchElementsUrl);
-
-    return backendFetchJson(fetchElementsUrl);
-}
-
-export function fetchVoltageLevelEquipments(
-    studyUuid,
-    currentNodeUuid,
-    substationsIds,
-    voltageLevelId,
-    inUpstreamBuiltParentNode
-) {
-    console.info(
-        `Fetching equipments of study '${studyUuid}' and node '${currentNodeUuid}' and voltage level '${voltageLevelId}' with substations ids '${substationsIds}'...`
-    );
-    let urlSearchParams = new URLSearchParams();
-    if (inUpstreamBuiltParentNode !== undefined) {
-        urlSearchParams.append(
-            'inUpstreamBuiltParentNode',
-            inUpstreamBuiltParentNode
-        );
-    }
-
-    const fetchEquipmentsUrl =
-        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
-        '/network-map' +
-        '/voltage-level-equipments/' +
-        encodeURIComponent(voltageLevelId) +
-        '?' +
-        getQueryParamsList(substationsIds, 'substationId') +
-        urlSearchParams.toString();
-    console.debug(fetchEquipmentsUrl);
-    return backendFetchJson(fetchEquipmentsUrl);
-}
-
-export function fetchEquipmentsIds(
-    studyUuid,
-    currentNodeUuid,
-    substationsIds,
-    equipmentType,
-    inUpstreamBuiltParentNode
-) {
-    console.info(
-        `Fetching equipments ids '${equipmentType}' of study '${studyUuid}' and node '${currentNodeUuid}' with substations ids '${substationsIds}'...`
-    );
-    let urlSearchParams = new URLSearchParams();
-
-    let fetchEquipmentsUrl =
-        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
-        '/network-map/' +
-        'equipments-ids' +
-        '?' +
-        'equipmentType=' +
-        equipmentType +
-        getQueryParamsList(substationsIds, 'substationId');
-
-    if (inUpstreamBuiltParentNode !== undefined) {
-        urlSearchParams.append(
-            'inUpstreamBuiltParentNode',
-            inUpstreamBuiltParentNode
-        );
-        fetchEquipmentsUrl =
-            fetchEquipmentsUrl + '&' + urlSearchParams.toString();
-    }
-    console.debug(fetchEquipmentsUrl);
-    return backendFetchJson(fetchEquipmentsUrl);
-}
-
-export function fetchLineOrTransformer(
-    studyUuid,
-    currentNodeUuid,
-    equipmentId
-) {
-    console.info(
-        `Fetching specific equipment '${equipmentId}' of type branch-or-3wt of study '${studyUuid}' and node '${currentNodeUuid}' ...`
-    );
-    let urlSearchParams = new URLSearchParams();
-    urlSearchParams.append('inUpstreamBuiltParentNode', true);
-    const fetchEquipmentInfosUrl =
-        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
-        '/network-map/branch-or-3wt/' +
-        encodeURIComponent(equipmentId) +
-        '?' +
-        urlSearchParams.toString();
-    console.debug(fetchEquipmentInfosUrl);
-    return backendFetchJson(fetchEquipmentInfosUrl);
-}
-
 export function fetchLimitViolations(
     studyUuid,
     currentNodeUuid,
@@ -716,51 +276,6 @@ export function fetchLimitViolations(
         '/limit-violations?limitReduction=' +
         limitReduction.toString();
     return backendFetchJson(url);
-}
-
-export function fetchBusesForVoltageLevel(
-    studyUuid,
-    currentNodeUuid,
-    voltageLevelId
-) {
-    console.info(
-        `Fetching buses of study '${studyUuid}' and node '${currentNodeUuid}' + ' for voltage level '${voltageLevelId}'...`
-    );
-    let urlSearchParams = new URLSearchParams();
-    urlSearchParams.append('inUpstreamBuiltParentNode', true);
-
-    const fetchBusesUrl =
-        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
-        '/network/voltage-levels/' +
-        encodeURIComponent(voltageLevelId) +
-        '/buses' +
-        '?' +
-        urlSearchParams.toString();
-    console.debug(fetchBusesUrl);
-    return backendFetchJson(fetchBusesUrl);
-}
-
-export function fetchBusbarSectionsForVoltageLevel(
-    studyUuid,
-    currentNodeUuid,
-    voltageLevelId
-) {
-    console.info(
-        `Fetching busbar sections of study '${studyUuid}' and node '${currentNodeUuid}' + ' for voltage level '${voltageLevelId}'...`
-    );
-    let urlSearchParams = new URLSearchParams();
-    urlSearchParams.append('inUpstreamBuiltParentNode', true);
-
-    const fetchBusbarSectionsUrl =
-        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
-        '/network/voltage-levels/' +
-        encodeURIComponent(voltageLevelId) +
-        '/busbar-sections' +
-        '?' +
-        urlSearchParams.toString();
-
-    console.debug(fetchBusbarSectionsUrl);
-    return backendFetchJson(fetchBusbarSectionsUrl);
 }
 
 export function fetchLinePositions(studyUuid, currentNodeUuid, linesIds) {
@@ -816,14 +331,6 @@ export function fetchNetworkModificationTree(studyUuid) {
     return backendFetchJson(url);
 }
 
-export function fetchNetworkModificationTreeNode(studyUuid, nodeUuid) {
-    console.info('Fetching network modification tree node : ', nodeUuid);
-    const url =
-        getStudyUrl(studyUuid) + '/tree/nodes/' + encodeURIComponent(nodeUuid);
-    console.debug(url);
-    return backendFetchJson(url);
-}
-
 export function fetchNetworkModificationSubtree(studyUuid, parentNodeUuid) {
     console.info('Fetching network modification tree node : ', parentNodeUuid);
     const url =
@@ -832,152 +339,6 @@ export function fetchNetworkModificationSubtree(studyUuid, parentNodeUuid) {
         encodeURIComponent(parentNodeUuid);
     console.debug(url);
     return backendFetchJson(url);
-}
-
-export function createTreeNode(studyUuid, parentId, insertMode, node) {
-    const nodeCreationUrl =
-        getStudyUrl(studyUuid) +
-        '/tree/nodes/' +
-        encodeURIComponent(parentId) +
-        '?mode=' +
-        insertMode;
-    console.debug('%s with body: %s', nodeCreationUrl, node);
-    return backendFetch(nodeCreationUrl, {
-        method: 'post',
-        headers: {
-            Accept: 'application/json',
-            'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(node),
-    });
-}
-
-export function deleteTreeNode(studyUuid, nodeId) {
-    console.info('Deleting tree node : ', nodeId);
-    const url =
-        getStudyUrl(studyUuid) + '/tree/nodes/' + encodeURIComponent(nodeId);
-    console.debug(url);
-    return backendFetch(url, {
-        method: 'delete',
-    });
-}
-
-export function deleteSubtree(studyUuid, parentNodeId) {
-    console.info('Deleting node subtree : ', parentNodeId);
-    const url =
-        getStudyUrl(studyUuid) +
-        '/tree/nodes/' +
-        encodeURIComponent(parentNodeId) +
-        '?deleteChildren=true';
-    console.debug(url);
-    return backendFetch(url, {
-        method: 'delete',
-    });
-}
-
-export function updateTreeNode(studyUuid, node) {
-    const nodeUpdateUrl = getStudyUrl(studyUuid) + '/tree/nodes/';
-    console.debug(nodeUpdateUrl);
-    return backendFetch(nodeUpdateUrl, {
-        method: 'put',
-        headers: {
-            Accept: 'application/json',
-            'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(node),
-    });
-}
-
-export function copyTreeNode(
-    sourceStudyUuid,
-    targetStudyId,
-    nodeToCopyUuid,
-    referenceNodeUuid,
-    insertMode
-) {
-    const nodeCopyUrl =
-        getStudyUrl(targetStudyId) +
-        '/tree/nodes?insertMode=' +
-        insertMode +
-        '&nodeToCopyUuid=' +
-        nodeToCopyUuid +
-        '&referenceNodeUuid=' +
-        referenceNodeUuid +
-        '&sourceStudyUuid=' +
-        sourceStudyUuid;
-    console.debug(nodeCopyUrl);
-    return backendFetch(nodeCopyUrl, {
-        method: 'post',
-        headers: {
-            Accept: 'application/json',
-            'Content-Type': 'application/json',
-        },
-    });
-}
-
-export function cutTreeNode(
-    studyUuid,
-    nodeToCutUuid,
-    referenceNodeUuid,
-    insertMode
-) {
-    const nodeCutUrl =
-        getStudyUrl(studyUuid) +
-        '/tree/nodes?insertMode=' +
-        insertMode +
-        '&nodeToCutUuid=' +
-        nodeToCutUuid +
-        '&referenceNodeUuid=' +
-        referenceNodeUuid;
-    console.debug(nodeCutUrl);
-    return backendFetch(nodeCutUrl, {
-        method: 'post',
-        headers: {
-            Accept: 'application/json',
-            'Content-Type': 'application/json',
-        },
-    });
-}
-
-export function cutSubtree(targetStudyId, nodeToCopyUuid, referenceNodeUuid) {
-    const nodeCopyUrl =
-        getStudyUrl(targetStudyId) +
-        '/tree/subtrees?subtreeToCutParentNodeUuid=' +
-        nodeToCopyUuid +
-        '&referenceNodeUuid=' +
-        referenceNodeUuid;
-    console.debug(nodeCopyUrl);
-    return backendFetch(nodeCopyUrl, {
-        method: 'post',
-        headers: {
-            Accept: 'application/json',
-            'Content-Type': 'application/json',
-        },
-    });
-}
-
-export function copySubtree(
-    sourceStudyUuid,
-    targetStudyUuid,
-    nodeToCopyUuid,
-    referenceNodeUuid
-) {
-    const nodeCopyUrl =
-        getStudyUrl(targetStudyUuid) +
-        '/tree/subtrees?subtreeToCopyParentNodeUuid=' +
-        nodeToCopyUuid +
-        '&referenceNodeUuid=' +
-        referenceNodeUuid +
-        '&sourceStudyUuid=' +
-        sourceStudyUuid;
-    console.debug(nodeCopyUrl);
-    return backendFetch(nodeCopyUrl, {
-        method: 'post',
-        headers: {
-            Accept: 'application/json',
-            'Content-Type': 'application/json',
-        },
-    });
 }
 
 export function copyOrMoveModifications(
@@ -1095,63 +456,4 @@ export function getUniqueNodeName(studyUuid) {
     const uniqueNodeNameUrl = getStudyUrl(studyUuid) + '/nodes/nextUniqueName';
     console.debug(uniqueNodeNameUrl);
     return backendFetchText(uniqueNodeNameUrl);
-}
-
-export function fetchSubstationsMapInfos(
-    studyUuid,
-    currentNodeUuid,
-    substationsIds,
-    inUpstreamBuiltParentNode
-) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.SUBSTATION.type,
-        EQUIPMENT_INFOS_TYPES.MAP.type,
-        inUpstreamBuiltParentNode
-    );
-}
-
-export function fetchLinesMapInfos(
-    studyUuid,
-    currentNodeUuid,
-    substationsIds,
-    inUpstreamBuiltParentNode
-) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.LINE.type,
-        EQUIPMENT_INFOS_TYPES.MAP.type,
-        inUpstreamBuiltParentNode
-    );
-}
-
-export function fetchHvdcLinesMapInfos(
-    studyUuid,
-    currentNodeUuid,
-    substationsIds,
-    inUpstreamBuiltParentNode
-) {
-    return fetchNetworkElementsInfos(
-        studyUuid,
-        currentNodeUuid,
-        substationsIds,
-        EQUIPMENT_TYPES.HVDC_LINE.type,
-        EQUIPMENT_INFOS_TYPES.MAP.type,
-        inUpstreamBuiltParentNode
-    );
-}
-
-export function cloneVoltageInitModifications(studyUuid, currentNodeId) {
-    console.info('cloning voltage init modifications');
-    const cloneVoltageInitModificationsUrl =
-        getStudyUrlWithNodeUuid(studyUuid, currentNodeId) +
-        '/voltage-init/modifications';
-
-    return backendFetch(cloneVoltageInitModificationsUrl, {
-        method: 'PUT',
-    });
 }
